refactured create_booking_screen.dart build method

+++ lib/repositories/booking_repository.dart
@@ -81,7 +81,7 @@ class BookingRepository {
     required int petId,
     required String date,
     required String time,
-    required double duration,
+    required int duration,
   }) async {
     final response = await supabaseClient
         .from('booking')
